requirements.txt
fastapi
uvicorn
PyPDF2
sentence-transformers
faiss-cpu
numpy
ollama
python-multipart
reportlab
